fix: Free LOB locators using CLI functions
Instead of allocating new handles, we can use the SQL_ATTR_FREE_LOCATOR
connection attribute to do this. Currently it is set to do one at a
time, but this attribute takes in an array of locators to free, with the
locator count in the last parameter. We could be more efficient by
queueing these locators to be freed at some later point, like at the end
of fetch and free them all at once.